backend/requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
coverage==7.13.5
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gTTS==2.5.4
gunicorn==25.3.0
idna==3.13
iniconfig==2.3.0
model-bakery==1.23.4
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-cov==7.1.0
pytest-django==4.12.0
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
google-generativeai
cryptography
python-json-logger
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
